requirements.txt
torch
transformers
accelerate
langchain
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
python-dotenv
unstructured
markdown
langchain_mistralai